integration: extract tracing.rs
